Split helper file into several partials
