+++ library/src/scripts/styles/styleHelpersSpinner.ts
@@ -1,0 +1,51 @@
+import { keyframes } from "typestyle";
+import { ColorHelper, deg, percent, quote } from "csx";
+import { globalVariables } from "@library/styles/globalStyleVars";
+import { ContentProperty, DisplayProperty, PositionProperty } from "csstype";
+import { debugHelper, unit } from "@library/styles/styleHelpers";
+
+const spinnerOffset = 73;
+const spinnerLoaderAnimation = keyframes({
+    "0%": { transform: `rotate(${deg(spinnerOffset)})` },
+    "100%": { transform: `rotate(${deg(360 + spinnerOffset)})` },
+});
+
+export interface ISpinnerProps {
+    color?: ColorHelper;
+    dimensions?: string | number;
+    thickness?: string | number;
+    size?: string | number;
+    speed?: string;
+}
+
+export const spinnerLoader = (props: ISpinnerProps) => {
+    const debug = debugHelper("spinnerLoader");
+    const globalVars = globalVariables();
+    const spinnerVars = {
+        color: props.color || globalVars.mainColors.primary,
+        size: props.size || 18,
+        thickness: props.thickness || 3,
+        speed: "0.7s",
+        ...props,
+    };
+    return {
+        ...debug.name("spinner"),
+        position: "relative" as PositionProperty,
+        content: quote("") as ContentProperty,
+        ...defaultTransition("opacity"),
+        display: "block" as DisplayProperty,
+        width: unit(spinnerVars.size),
+        height: unit(spinnerVars.size),
+        borderRadius: percent(50),
+        borderTop: `${unit(spinnerVars.thickness)} solid ${spinnerVars.color.toString()}`,
+        borderRight: `${unit(spinnerVars.thickness)} solid ${spinnerVars.color.fade(0.3).toString()}`,
+        borderBottom: `${unit(spinnerVars.thickness)} solid ${spinnerVars.color.fade(0.3).toString()}`,
+        borderLeft: `${unit(spinnerVars.thickness)} solid ${spinnerVars.color.fade(0.3).toString()}`,
+        transform: "translateZ(0)",
+        animation: `spillerLoader ${spinnerVars.speed} infinite ease-in-out`,
+        animationName: spinnerLoaderAnimation,
+        animationDuration: spinnerVars.speed,
+        animationIterationCount: "infinite",
+        animationTimingFunction: "ease-in-out",
+    };
+};
